+++ internal/render/render.go
@@ -2,66 +2,66 @@ package render
 
 import (
 	"errors"
-	"geoserver/internal/db/models"
-	"image"
-	"os"
+	"fmt"
+	"geoserver/internal/translator"
+	"math"
+	"os/exec"
 	"strconv"
-	"sync"
-	"time"
+
+	"github.com/Lvov-SA/gdal"
 )
 
-func Tiler(layer models.Layer, z, x, y int) (image.Image, error) {
-	filePath := "../resource/cache/" + layer.Name + "/" + strconv.Itoa(z) + "/"
-	fileName := strconv.Itoa(x) + "_" + strconv.Itoa(y) + ".png"
-	file, err := os.Open(filePath + fileName)
-	if os.IsNotExist(err) {
-		resultChan := make(chan Result, 1)
-		Tasks <- Task{
-			layer:    layer,
-			filePath: filePath,
-			fileName: fileName,
-			x:        x,
-			y:        y,
-			z:        z,
-			result:   resultChan,
-			wg:       nil,
-		}
-		select {
-		case result := <-resultChan:
-			{
-				if result.err != nil {
-					return nil, err
-				}
-			}
-		case <-time.After(10 * time.Minute):
-			return nil, errors.New("Истекло время ожидания воркера")
-		}
-		file, err = os.Open(filePath + fileName)
-		if err != nil {
-			return nil, err
-		}
+func CliWarpRender(task Task) error {
+	minX, minY, maxX, maxY := translator.WebMercarator(task.x, task.y, task.z)
+
+	args := []string{
+		"-s_srs", task.layer.Projection,
+		"-t_srs", "EPSG:3857",
+		"-te",
+		fmt.Sprintf("%f", minX),
+		fmt.Sprintf("%f", minY),
+		fmt.Sprintf("%f", maxX),
+		fmt.Sprintf("%f", maxY),
+		"-ts",
+		strconv.Itoa(task.layer.TileSize),
+		strconv.Itoa(task.layer.TileSize),
+		"-r", "near",
+		"-of", "PNG",
+		"-co", "COMPRESS=DEFLATE",
+		"-co", "ZLEVEL=6",
+		"-overwrite",
+		"../resource/map/" + task.layer.SourcePath,
+		task.filePath + task.fileName,
 	}
 
-	imageRGBA, _, err := image.Decode(file)
-	if err != nil {
-		file.Close()
-		return nil, err
-	}
-	file.Close()
-	return imageRGBA, nil
+	cmd := exec.Command("gdalwarp", args...)
+	err := cmd.Run()
+	return err
 }
 
-func MakeTask(layer models.Layer, z, x, y int, resultChan chan Result, wg *sync.WaitGroup) {
-	filePath := "../resource/cache/" + layer.Name + "/" + strconv.Itoa(z) + "/"
-	fileName := strconv.Itoa(x) + "_" + strconv.Itoa(y) + ".png"
-	Tasks <- Task{
-		layer:    layer,
-		filePath: filePath,
-		fileName: fileName,
-		x:        x,
-		y:        y,
-		z:        z,
-		result:   resultChan,
-		wg:       wg,
+func TranslateRender(task Task) error {
+	coef := math.Pow(2, float64(task.z))
+	maxSize := min(task.layer.Width, task.layer.Height)
+	xFloat := float64(task.x)
+	yFloat := float64(task.y)
+	maxSizeFloat := float64(maxSize)
+	readSize := maxSizeFloat / coef
+	if xFloat*readSize >= float64(task.layer.Width) || yFloat*readSize >= float64(task.layer.Height) {
+
+		return errors.New("Выход за границы")
 	}
+	options := []string{"-srcwin",
+		fmt.Sprintf("%d", int(xFloat*readSize)),
+		fmt.Sprintf("%d", int(yFloat*readSize)),
+		fmt.Sprintf("%d", int(readSize)),
+		fmt.Sprintf("%d", int(readSize)),
+		"-outsize",
+		fmt.Sprintf("%d", task.layer.TileSize),
+		fmt.Sprintf("%d", task.layer.TileSize)}
+	err := gdal.ConvertTile(
+		"../resource/map/"+task.layer.SourcePath,
+		task.filePath+task.fileName,
+		options,
+	)
+	return err
 }
